third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFECompositeElement-svgdom-operator-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'operator' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_IN
PASS overComposite1.operator.baseVal is SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ARITHMETIC
PASS successfullyParsed is true

TEST COMPLETE
